Upgraded layout of feed
The feed now displays in a 2 column layout with feed items on the left
and feed content on the right similar to modern email clients

+++ app/blueprints/feed/views.py
@@ -1,11 +1,15 @@
 """Feed Routes"""
 
+import logging
+
 from flask import render_template, request
 from flask_login import login_required, current_user
 
 from app.services import subscription as subscription_service
 from app.services import post as post_service
 from .forms import SubscriptionForm
+
+log = logging.getLogger(__name__)
 
 @login_required
 def feed():
@@ -23,20 +27,51 @@ def feed():
     return render_template('feed/feed.html', **data)
 
 @login_required
+def feed_list():
+    """Load feed list"""
+    page = int(request.args.get('page', '1'))
+    per_page = 20
+    saved = request.args.get('saved', '').lower() == 'true'
+    if saved:
+        posts = post_service.get_saved(current_user, page=page, per_page=per_page)
+    else:
+        posts = post_service.get_all(current_user, page=page, per_page=per_page)
+    data = {
+        'posts': posts
+    }
+    return render_template('feed/feed-list.html', **data)
+
+@login_required
+def post_content(post_id):
+    """Render content for single post"""
+    # TODO turn this into a class based view
+    if request.method == 'GET': # return post content
+        post = post_service.get_one(post_id)
+        data = {
+            'post': post,
+        }
+        return render_template('feed/post-content.html', **data)
+    elif request.method == 'DELETE': # delete post
+        post_service.delete(post_id)
+        return '', 200
+
+@login_required
 def toggle_save_post(post_id):
     """Updated saved status for post"""
+    target_id = request.headers.get('HX-Target')
     post = post_service.toggle_save(post_id)
     data = {
         'post': post,
     }
-    return render_template('feed/post.html', **data)
+    # check which element triggered the request, adjust template accordingly
+    log.debug('Target ID %s', target_id)
+    if target_id == 'post-{}'.format(post.id):
+        template = 'feed/post.html'
+    else:
+        template = 'feed/post-content.html'
 
-@login_required
-def delete_post(post_id):
-    """Delete post"""
-    post_service.delete(post_id)
-    return '', 200
-    
+    return render_template(template, **data)
+
 @login_required
 def subscriptions():
     """Load subscriptions"""
